+++ NEWS.md
@@ -1,9 +1,15 @@
 
+
+# archeofrag.gui 1.1.4
+
+* In 'Spatial units optimisation', add mean and standard deviation for admixture. 
+
 # archeofrag.gui 1.1.3
 Released: 2025-
 
 * Replace the admixture table in 'Measurements' tab by a dissimilarity matrix and add a download button. Add Baker's Gamme index. Improve entanglement computation.
 * In the 'Measurements' tab, add new function to rank the spatial units based on their relative cohesion values.   
+* Small changes in the openMOLE code generator. 
 
 # archeofrag.gui 1.1.2
 Released: 2025-05-27